+++ test/lang/en/complex/compound/minute-span-across-hour-range.js
@@ -15,7 +15,7 @@ describe('Minute span across an hour range:', function() {
   describe('minute range', function() {
     run([
       ['0-30 9-17 * * *',
-        'every minute from zero through 30 past the hour, ' +
+        'every minute from 0 through 30 past the hour, ' +
         'from 9 a.m. through 5:30 p.m.']
     ]);
   });
@@ -24,7 +24,7 @@ describe('Minute span across an hour range:', function() {
     run([
       ['* 9-17 * * MON', 'every minute from 9 a.m. through 5:59 p.m. on Monday'],
       ['0-30 9-17 * * MON-FRI',
-        'every minute from zero through 30 past the hour, ' +
+        'every minute from 0 through 30 past the hour, ' +
         'from 9 a.m. through 5:30 p.m. on Monday through Friday']
     ]);
   });