+++ test/lang/en/complex/ranges/objects.js
@@ -2,9 +2,9 @@ import {run} from '../../../../runner.js';
 
 describe('Valid objects with ranges:', function() {
   var tests = [
-    [{second: '0-30'}, 'every second from zero through 30 past the minute'],
-    [{second: '10-20'}, 'every second from ten through 20 past the minute'],
-    [{minute: '0-29'}, 'every minute from zero through 29 past the hour'],
+    [{second: '0-30'}, 'every second from 0 through 30 past the minute'],
+    [{second: '10-20'}, 'every second from 10 through 20 past the minute'],
+    [{minute: '0-29'}, 'every minute from 0 through 29 past the hour'],
     [{minute: '1-5'}, 'every minute from one through five past the hour'],
     [{hour: '9-17'}, 'every hour from 9 a.m. through 5 p.m.'],
     [{hour: '0-5'}, 'every hour from midnight through 5 a.m.'],